requirements.txt
python-telegram-bot==13.15
requests==2.31.0
numpy==1.26.0
pandas==2.2.0
schedule==1.2.0
